% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{compBoxplot}
\alias{compBoxplot}
\title{Compare via boxplots}
\usage{
compBoxplot(par, df)
}
\arguments{
\item{par}{A character string representing the parameter to be plotted.}

\item{df}{A data frame containing the data for plotting.}
}
\value{
A ggplot2 plot.
}
\description{
This function creates a boxplot to compare a parameter between groups.
}
\examples{
par <- "example_par"
df <- data.frame(
    group = c("group1", "group1", "group1", "group2", "group2", "group2"),
    example_par = c(3, 7, 8, 5, 9, 6)
)
compBoxplot(par, df)
}
