% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_category}
\alias{count_category}
\title{Count categories}
\usage{
count_category(data, category)
}
\arguments{
\item{data}{data frame}

\item{category}{character string representing the category}
}
\value{
save categories to folder \verb{/analysis/relative/categories/}
}
\description{
count categories and save as csv
}
\examples{
\dontrun{
lapply(sel_categories, count_category, data = combined_complete)
}
}
