% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_distance_fun}
\alias{date_distance_fun}
\title{Calculate the distance between two dates}
\usage{
date_distance_fun(v1, v2, max_dist = 1)
}
\arguments{
\item{v1}{first date}

\item{v2}{second date}

\item{max_dist}{maximum distance in days between the two dates}
}
\value{
a data frame with a single column include indicating
whether the distance is within the maximum distance
}
\description{
Calculate the absolute difference between two dates in days
}
