% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_category}
\alias{plot_category}
\title{Plot Category}
\usage{
plot_category(data, category, width, height, output_dir = NULL)
}
\arguments{
\item{data}{A data frame containing the category column.}

\item{category}{The name of the category column in the data frame.}

\item{width}{The width of the plot.}

\item{height}{The height of the plot.}

\item{output_dir}{The directory to save the plot. If NULL, the plot saved to}
}
\value{
A ggplot2 plot object.
}
\description{
This function takes a data frame and a category column as input,
and returns a ggplot2 plot of the category.
}
\details{
This function plots a category using ggplot2.
}
\examples{
\dontrun{
plot_category(data = combined_complete, category = "dx_icd_level1", width = 4, height = 2)
}
}
