% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotConfMat}
\alias{plotConfMat}
\title{Plot confusion matrix}
\usage{
plotConfMat(last_fit, name, metric_df, output_dir)
}
\arguments{
\item{last_fit}{The last fit from the tuning process.}

\item{name}{A character string of the name of the model.}

\item{metric_df}{A data frame containing the metric}

\item{output_dir}{The directory to save the plot.}
}
\value{
A ggplot2 object.
}
\description{
This function plots the confusion matrix from a tuned
xgboost model.
}
\examples{
set.seed(12)
mtcars$cluster <- as.character(mtcars$am)
folds <- rsample::vfold_cv(mtcars, v = 2)
rec <- recipes::recipe(cluster ~ ., data = mtcars)
mod <- parsnip::logistic_reg()
control <- tune::control_resamples(save_pred = TRUE)
metrics <- yardstick::metric_set(
    yardstick::accuracy,
    yardstick::bal_accuracy,
    yardstick::f_meas,
    yardstick::roc_auc
)
res <- tune::fit_resamples(mod, rec, folds, control = control, metrics = metrics)
metric_df <- tune::collect_metrics(res)
metric_df$.estimate <- metric_df$mean
plotConfMat(res, name = "model", output_dir = tempdir(), metric_df = metric_df)
}
