% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_this}
\alias{scale_this}
\title{Scale a variable}
\usage{
scale_this(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric vector with the same length as x
}
\description{
Scale a variable by subtracting the mean and dividing by the standard deviation
}
\examples{
scale_this(1:10)
}
