% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{stabilityBlood}
\alias{stabilityBlood}
\title{Stability metric for Blood}
\usage{
stabilityBlood(t, df, vars_cont, normal_estimate, ndim)
}
\arguments{
\item{t}{The seed number for reproducibility.}

\item{df}{The data frame containing the data for testing.}

\item{vars_cont}{Continuous variables for datathin.}

\item{normal_estimate}{Estimates for normally distributed parameters for datathin.}

\item{ndim}{Number of dimensions for Seurat.}
}
\value{
A named vector of stability values.
}
\description{
This function calculates the stability of Seurat clustering
at different resolutions for blood data.
}
\examples{
set.seed(1)
df <- data.frame(
    a = stats::rnorm(100, mean = 2, sd = 1),
    b = stats::rnorm(100, mean = 2, sd = 2),
    c = stats::rnorm(100, mean = 2, sd = 1),
    d = stats::rnorm(100, mean = 2, sd = 2),
    e = stats::rnorm(100, mean = 2, sd = 1)
)
vars_cont <- c("a", "b", "c", "d", "e")
normal_estimate <-
    matrix(c(0.5, 0.1, 0.7, 0.2, 0.3),
        nrow = nrow(df),
        ncol = length(vars_cont)
    )
ndim <- 2
stabilityBlood(
    t = 1,
    df = df,
    vars_cont = vars_cont,
    normal_estimate = normal_estimate,
    ndim = ndim
)
}
