% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{stabilityCSF}
\alias{stabilityCSF}
\title{Stability metric for CSF}
\usage{
stabilityCSF(
  t,
  df,
  vars_cont,
  vars_cat,
  normal_estimate,
  weibull_estimate,
  ndim
)
}
\arguments{
\item{t}{The seed number for reproducibility.}

\item{df}{The data frame containing the data for testing.}

\item{vars_cont}{Continuous variables for datathin.}

\item{vars_cat}{Categorical variables for datathin.}

\item{normal_estimate}{Estimates for normally distributed parameters for datathin.}

\item{weibull_estimate}{Estimates for weibull distributed parameters for datathin.}

\item{ndim}{Number of dimensions for Seurat}
}
\value{
A named vector of stability values.
}
\description{
This function calculates the stability of Seurat clustering
at different resolutions.
}
\examples{
set.seed(1)
df <- data.frame(
    a = rnorm(100, mean = 2, sd = 1),
    b = rnorm(100, mean = 2, sd = 1),
    x = rpois(100, lambda = 2),
    y = rpois(100, lambda = 2),
    z = rpois(100, lambda = 2)
)
vars_cont <- c("a", "b")
vars_cat <- c("x", "y", "z")
normal_estimate <-
    matrix(c(0.5, 0.2),
        nrow = nrow(df),
        ncol = length(vars_cont)
    )
weibull_estimate <-
    matrix(c(0.5, 0.2, 0.7),
        nrow = nrow(df),
        ncol = length(vars_cat)
    )
ndim <- 2
stabilityCSF(
     t = 1,
     df = df,
     vars_cont = vars_cont,
     vars_cat = vars_cat,
     normal_estimate = normal_estimate,
     weibull_estimate = weibull_estimate,
     ndim = ndim
)
}
