% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{timePlot}
\alias{timePlot}
\title{Time Plot}
\usage{
timePlot(data, var, size, span)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{var}{A character string representing the variable to be plotted.}

\item{size}{A numeric value representing the size of the points in the plot.}

\item{span}{A numeric value representing the span for the loess smoothing.}
}
\value{
A ggplot2 object representing the time plot.
}
\description{
This function creates a time plot for a given variable in the data.
}
\examples{
data <- data.frame(
    measure_time = seq(1, 100, by = 1),
    var1 = rnorm(100),
    var2 = rnorm(100)
)
timePlot(data, var = "var1", size = 1, span = 0.5)
}
