% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{abundanceCategoryPlot}
\alias{abundanceCategoryPlot}
\title{the abundance of each gene in a cluster as a barplot.}
\usage{
abundanceCategoryPlot(data, cluster, output_dir = NULL)
}
\arguments{
\item{data}{dataframe with the gene names, q-values and TF-IDF values.}

\item{cluster}{the cluster to plot.}

\item{output_dir}{the directory to save the plot. If NULL, the plot is saved to \verb{/analysis/relative/abundance/}.}
}
\value{
a ggplot2 plot saved to output_dir
}
\description{
the abundance of each gene in a cluster as a barplot.
}
\examples{
\dontrun{
lapply(unique(seu_csf_train$cluster), abundanceCategoryPlot, data = abundance_combined_soupx_csf_norm_datathin)
}
}
