% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{corrPlot}
\alias{corrPlot}
\title{Individual correlation plots of variables}
\usage{
corrPlot(var, estimate_df, plot_df, output_dir)
}
\arguments{
\item{var}{character string representing the variable}

\item{estimate_df}{data frame containing the estimates}

\item{plot_df}{data frame containing the plot}

\item{output_dir}{character string representing the directory to save the
plot.}
}
\value{
a ggplot2 plot saved to output_dir
}
\description{
Create individual correlation plots of selected variables
}
\examples{
var <- "example_var"
estimate_df <- data.frame(
    var = c("example_var", "example_var", "example_var"),
    age = c(20, 30, 40),
    estimate = c(0.5, 0.6, 0.7),
    p_adjust = c(0.01, 0.02, 0.03)
)
plot_df <- data.frame(
    age = c(20, 30, 40),
    example_var = c(1, 2, 3)
)
output_dir <- "."
corrPlot(var, estimate_df, plot_df, output_dir)
}
