% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{corrPlot}
\alias{corrPlot}
\title{Individual correlation plots of variables}
\usage{
corrPlot(var, estimate_df, plot_df, output_dir = NULL)
}
\arguments{
\item{var}{character string representing the variable}

\item{estimate_df}{data frame containing the estimates}

\item{plot_df}{data frame containing the plot}

\item{output_dir}{character string representing the directory to save the
plot. If NULL, the plot is saved to \verb{/analysis/relative/relative/}.}
}
\value{
a ggplot2 plot saved to output_dir
}
\description{
Create individual correlation plots of selected variables
}
\examples{
\dontrun{
lapply(age_var, corrPlot)
}
}
