% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_category}
\alias{count_category}
\title{Count categories}
\usage{
count_category(data, category, output_dir)
}
\arguments{
\item{data}{data frame}

\item{category}{character string representing the category}

\item{output_dir}{character string representing the output directory}
}
\value{
save categories to folder \verb{/analysis/relative/categories/}
}
\description{
count categories and save as csv
}
\examples{
df <- data.frame(
  category = c("A", "B", "A", "C", "B", "A"),
  value = 1:6
)

temp_dir <- tempdir()

count_category(df, "category", temp_dir)

}
