% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{heatmap_group_csf}
\alias{heatmap_group_csf}
\title{Grouped heatmap}
\usage{
heatmap_group_csf(
  category,
  data,
  label,
  cutree_rows,
  height,
  transform = FALSE,
  cutree_cols = 8,
  output_dir = NULL
)
}
\arguments{
\item{category}{character string representing the category}

\item{data}{data frame}

\item{label}{character string representing the label}

\item{cutree_rows}{numeric value representing the number of clusters the rows are divied into}

\item{height}{numeric value representing the height of the heatmap}

\item{transform}{boolean value representing if the data should be transposed}

\item{cutree_cols}{numeric value representing the number of clusters the columns are divied into, default: 8}

\item{output_dir}{character string representing the directory to save the heatmap, if NULL, the heatmap is saved to \verb{/analysis/relative/heatmap/}}
}
\value{
save grouped heatmap to folder
}
\description{
Plot goruped heatmap and save to folder
}
\examples{
\dontrun{
heatmap_group_csf(category = "dx_icd_level1", data = csf_data, label = "CSF", cutree_rows = 4, height = 5)
}
}
