% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_category}
\alias{plot_category}
\title{Plot Category}
\usage{
plot_category(data, category, width, height, output_dir)
}
\arguments{
\item{data}{A data frame containing the category column.}

\item{category}{The name of the category column in the data frame.}

\item{width}{The width of the plot.}

\item{height}{The height of the plot.}

\item{output_dir}{The directory to save the plot.}
}
\value{
A ggplot2 plot object.
}
\description{
This function takes a data frame and a category column as input,
and returns a ggplot2 plot of the category.
}
\details{
This function plots a category using ggplot2.
}
\examples{
data <- data.frame(
  category = c("A", "A", "B", "B", "C", "C"),
  value = c(1, 2, 3, 4, 5, 6)
)

category <- "category"
# Call the plot_category function
plot_category(data, category, width = 5, height = 5, output_dir = ".")

}
